Change size_t to uint32_t

+++ src/bindings/js/node/src/helper.cpp
@@ -322,7 +322,7 @@ void fill_tensor_from_strings(ov::Tensor& tensor, const Napi::Array& arr) {
         OPENVINO_THROW("Passed array must have the same size (number of elements) as the Tensor!");
     }
     const auto data = tensor.data<std::string>();
-    for (size_t i = 0; i < tensor.get_size(); ++i) {
+    for (uint32_t i = 0; i < tensor.get_size(); ++i) {
         OPENVINO_ASSERT(arr[i].IsString(), "The array passed to create string tensor must contain only strings.");
         data[i] = arr[i].ToString().Utf8Value();
     }
